requirements.txt
flask==3.0.3
requests==2.32.3
pytz==2024.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
python-dateutil==2.9.0.post0
